make `$safe` variable argument
`if block` in L.96 seems not to be called never. It is because `$safe` variable is settled as `true` before the if block.
when move `$safe` to the function arguments, user can handle the `$safe` variable, and the `if block` will work!

In this PR,  the variable `$safe` moved to the function argument.